requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
ez-setup==0.9
psycopg2-binary==2.9.5
pytz==2022.6
sqlparse==0.4.3
